requirements.txt
streamlit
pandas
openpyxl
python-dotenv
gitpython
imapclient
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
psycopg2-binary
xlrd
lxml
requests